requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.10.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Pillow==11.1.0
python-dotenv==1.0.1
email-validator==2.1.1
Werkzeug==3.0.1 